Add toggle for ImGui NavEnableKeyboard flag
This flag is used to enable enhanced keyboard navigation for ImGui menus such
as using the tab key or arrow keys to swap between interactible UI elements, and
enter/space to interact with the currently selected UI element.

When this flag is disabled, tab navigation only works between text fields.

Unfortunately this toggle is currently mostly non-functional, as the enhanced
keyboard navigation behavior was added in ImGui 1.90, and Dalamud is currently
based on ImGui 1.88.